add snapshot test for copy page endpoint
